cloud_functions/get_cloudsql_instance_inventory/requirements.txt
google-cloud-asset==3.14.2
google-cloud-pubsub==2.13.10
google-cloud-logging==3.2.5
functions-framework==3.2.0